new conversion rules for task names
A task name is a simple string like "request:POST-DELETE" or
"http:avatar". The job processor splits the name at every colon and
converts each part into a constant in the module "Jobs".
"request:POST-DELETE" is hence interpreted as the class
"Jobs::Request::POST_DELETE".

If the class cannot be found a NameError is thrown (and handled by
printing an error message to the log). This should be avoided by passing
the job-processing script the name of a queue that only contains jobs it
can understand.