implement XMPP relay
The server accepts XMPP stanzas on a socket, transmits them as the XMPP
server component on successful parse and relays the response to the
sender. This interface is to be used by the job processor.

As the inputs are not checked, socket permissions should be set such
that only trusted processes can relay messages.