load private key to generate public key
We need the public key in a few responders to compare whether the provided
public key indicates the local server or a remote.  Eventually, we should
simply use domain names.

As the public key is stored in the Server module, we have to access it through
the static conf method.