Makefile to build deb/rpm packages
FPM can build rpm and deb packages from our gem dependencies.
FPM should be installed with:

  gem install fpm

We collect all dependent gems with bundler:

  bundle install --without development
  bundle pack

As the backend depends on libertree-client and libertree-model, quite a few
of the dependent gems are already covered by the Makefiles of these projects.

The dependent gem packages are configured to install to /usr/share/libertree/gems;
the backend libraries, config files and scripts will be installed to
/usr/share/libertree/backend-rb .