wait longer for response from XMPP server
When we are too impatient when reading from the socket, the socket will
be written to after we stopped listening.  The next time we listen on the
socket again, the previous message that we didn't wait for will already be
in the socket buffer.

Timing out a socket read is only a precaution for the case where the XMPP
server itself is not responding. The XMPP server will time out connections
to other servers automatically (Prosody has a default timeout of 60 seconds)
and write an error stanza to the socket.

It is preferrable to read the XMPP server's error stanza from the stream,
because eventually the XMPP server will push its error stanza to the socket
causing the request *immediately after* a timeout to fail invariably.