feed the parser char by char
We may not feed the whole chunk to the parser at once.
As soon as the parser reaches the end of the stanza it will
implode and a new parser will be used, discarding whatever
else is in the queue.  If the parser sees that there are chars
in the queue after closing off the stanza, it will throw an
error and the excess chars won't be processed.