move remaining two frontend jobs here
The jobs "http:avatar" (fetch remote members' avatars asynchronously) and
"http:embed" (fetch object embed snippet from media provider) are the last
remaining jobs that are handled by the separate frontend job queue.

As the backend job processor not only takes care of server-to-server
requests but also simply delayed tasks, it makes sense to fuse the two
and only leave a single job processor in the backend for tasks that are
to be executed in the background.