simple XMPP API for dumb clients
Messages sent to the gateway's JID by users with registered JIDs are
treated as commands to submit a new post or comment if they match the
following syntax:

  POST This is a post.

  REPLY 1234 This is a comment on the post with id 1234.

This may be useful for clients that don't have native PubSub support.
In a future extension of this feature any new post or comment (or only
those mentioned in notifications) on Libertree could be forwarded as
messages to the user and be displayed with post id.  This would allow
the user to discover the post id needed for replies more naturally.